More work on client/server export service
